Player1 Camera switches on boundary
Now, when player 1 moves off the edge of the screen, the camera will
switch the the camera in the area that he's in.

When we multiple players, I will do so with the other players, which
shouldn't be difficult.

When we get proc.gen maps, we should be able to code the coordinated of
each room to communicate with the CameraSelect functions for each player
to decide which camera to use. Not sure how quite yet though.