Camera fun
trying to make the cameras part of the player prefab, and have them,
based on PLAYERNUM, follow the proper character and render to the same
viewport.

Known bug:
If there is one sage, it works perfectly, and the camera follows the
player, and renders to 1/4 of the screen, based on the PLAYERNUM
(doesn't matter which, as long as it's 1-4)

But if there is multiple sages, it breaks, and you get on camera
rendering to the WHOLE screen, and you can only see the one.