Working splitscreen w/ HUD
So I have the players, their GUIs, and their cameras all set to prefabs,
so there are 4 sets, labeled accordingly.

I can't think of a better way to make this work until we solve the issue
of the multiple instances of the camera prefab breaking the splitscreen
effect.

controls and huds seem to be working well.