Hud completions
stuff got done... no idea why in the fuck it says 339 files to be
committted...

As far as I can tell, looking around, everything seems to work and be in
place. So for some reason, it deleted and remade a pile of stuff in the
same place.

If this breaks something, we can always revert it.