Test before I continue please
I've made a set path, I've linked the doors for the first 3 rooms in the
path. Would someone throw in a character and do a quick test for me
whilst I continue linking doors? The doors of the origin room and the
later made rooms are currently set to an invis sprite. The mini platform
on the left side of origin room has the only door currently, since we
only have caves.